requirements.txt
streamlit
ultralytics
Pillow
torch
numpy